moved the lbr hardware types
