+++ lbr_state_broadcaster/package.xml
@@ -16,6 +16,7 @@
   <build_depend>pluginlib</build_depend>
   <build_depend>lbr_state_msgs</build_depend>
   <build_depend>hardware_interface</build_depend>
+  <build_depend>lbr_hardware</build_depend>
 
   <exec_depend>rclcpp</exec_depend>
   <exec_depend>rclcpp_lifecycle</exec_depend>
@@ -24,6 +25,7 @@
   <exec_depend>pluginlib</exec_depend>
   <exec_depend>lbr_state_msgs</exec_depend>
   <exec_depend>hardware_interface</exec_depend>
+  <exec_depend>lbr_hardware</exec_depend>
 
   <test_depend>ament_cmake_gmock</test_depend>
   <test_depend>controller_manager</test_depend>
